Add files via upload
Long datatype added to PK's

+++ src/main/java/TheatreTicketBookingSystem/factories/EconomyFactory.java
@@ -10,7 +10,7 @@ import java.util.Map;
 public class EconomyFactory {
     public static Economy getEconomy(Map<String, Object> values){
         Economy factoryEconomy = new Economy.Builder()
-                .seat_id((Integer)values.get("seat_id"))
+                .seat_id((Long)values.get("seat_id"))
                 .price((Integer)values.get("price"))
                 .class_name((String)values.get("class_name"))
                 .build();
