+++ src/main/java/TheatreTicketBookingSystem/factories/ShowFactory.java
@@ -11,7 +11,7 @@ import java.util.Map;
 public class ShowFactory {
     public static Show getShow(Map<String, Object> values){
         Show factoryShow = new Show.Builder()
-                .show_id((Integer)values.get("show_id"))
+                .show_id((Long)values.get("show_id"))
                 .movie_id((Integer)values.get("movie_id"))
                 .time((String)values.get("time"))
                 .show_date((Date) values.get("show_date"))
